requirements.txt
groq
python-dotenv
streamlit
firebase-admin
requests
fastapi
uvicorn
pydantic